backend/requirements.txt
fastapi[standard]==0.128.0
feedparser==6.0.12
PyYAML==6.0.3
transformers==4.44.0
torch==2.10.0
spacy==3.8.11
